App Restart
Code exists in the that relauches the app without erasing all user data. This allows the app to keep the user signed in. Currently there is a test function in progress for testings the functionality of staying signed in. With this commit the other tests in the SignInTest.java file may not work as there is no method of pausing the tests while appium restarts the server and the app